+++ app/src/main/java/com/devstart/hymn/data/model/Response.kt
@@ -1,4 +1,4 @@
-package com.devstart.hymn.model
+package com.devstart.hymn.data.model
 
 import com.squareup.moshi.Json
