+++ app/src/main/java/com/devstart/hymn/splash/SplashScreenActivity.kt
@@ -1,9 +1,11 @@
-package com.devstart.hymn
+package com.devstart.hymn.splash
 
 import android.content.Intent
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import android.os.Handler
+import com.devstart.hymn.R
+import com.devstart.hymn.home.ui.MainActivity
 
 class SplashScreenActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
