+++ app/src/main/res/layout/activity_hymn_detail.xml
@@ -8,7 +8,7 @@
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".HymnDetailActivity">
+        tools:context=".hymn_detail.ui.HymnDetailActivity">
 
         <TextView
             android:id="@+id/textView"
