+++ app/src/main/res/layout/activity_splash_screen.xml
@@ -8,7 +8,7 @@
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".SplashScreenActivity">
+        tools:context=".splash.SplashScreenActivity">
 
         <ImageView
             android:id="@+id/imageView2"
